+++ js/updater.js
@@ -1,16 +1,17 @@
 // set version
-const wledGuiVersion = "0.6.1";
+const wledGuiVersion = require('electron').remote.app.getVersion();
 
-checkForUpdate();
+if (localStorage.getItem("remindLaterTime") === null || (Date.now() - localStorage.getItem("remindLaterTime")) >= 259200000) {  // 3 days
+    checkForUpdate();
+}
 
 // checks if a update is available
 function checkForUpdate() {
     let xhr = new XMLHttpRequest();
     xhr.open('GET', 'https://raw.githubusercontent.com/WoodyLetsCode/WLED-GUI/master/VERSION', true);
     xhr.onload = function () {
-        console.log();
-        if (xhr.response !== wledGuiVersion && (localStorage.getItem("remindLaterTime") === null || (Date.now() - localStorage.getItem("remindLaterTime")) >= 259200000)) { // 3 days
-            console.log("New update avaiable!");
+        if (xhr.response !== wledGuiVersion) {
+            log.info("New update avaiable!");
             let instance = M.Modal.getInstance(document.getElementById("updatePopup"));
             document.getElementById("updatePopupText").innerText = "A new update for WLED-GUI is available.\n\nYour version: " + wledGuiVersion + "\nLatest version: " + xhr.response;
             instance.open();