+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # WLED-GUI changelog
 
+## WLED-GUI version 0.7.0
+- added settings page
+- added autostart option
+- added autostart hidden option
+- added option to turn on lights automatically at startup
+- the window will now show when the renderer process has rendered the page (no visual flash anymore)
+- added tray icon option
+- added log file
+- added option to turn on lights automatically only at autostart
+- fixed #6
+
 ## WLED-GUI version 0.6.0
 - added an update notifier
 - added new device discovery method